@@ -8,53 +8,61 @@
     "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e",
+    "generate-barrels": "barrelsby -D -e spec.ts$ -d ./src/app/components && barrelsby -D -e spec.ts$ -d ./src/app/services && barrelsby -D -e spec.ts$ -d ./src/app/pipes"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.1.1",
-    "@angular/common": "^5.1.1",
-    "@angular/compiler": "^5.1.1",
-    "@angular/core": "^5.1.1",
-    "@angular/forms": "^5.1.1",
-    "@angular/http": "^5.1.1",
-    "@angular/platform-browser": "^5.1.1",
-    "@angular/platform-browser-dynamic": "^5.1.1",
-    "@angular/router": "^5.1.1",
-    "@types/lodash": "^4.14.91",
-    "bootstrap": "^3.3.7",
+    "@angular/animations": "^7.2.8",
+    "@angular/common": "^7.2.8",
+    "@angular/compiler": "^7.2.8",
+    "@angular/core": "^7.2.8",
+    "@angular/forms": "^7.2.8",
+    "@angular/http": "^7.2.8",
+    "@angular/platform-browser": "^7.2.8",
+    "@angular/platform-browser-dynamic": "^7.2.8",
+    "@angular/router": "^7.2.8",
+    "@types/lodash": "^4.14.107",
+    "bootstrap": "^4.3.1",
     "classlist.js": "^1.1.20150312",
     "compass-mixins": "^0.12.10",
-    "core-js": "^2.5.3",
+    "core-js": "^2.5.5",
     "intl": "^1.2.5",
-    "jquery": "1.11.3",
-    "lodash": "^4.17.4",
-    "ng-recaptcha": "^3.0.2",
-    "ngx-bootstrap": "^2.0.0-rc.0",
-    "rxjs": "^5.5.5",
+    "jquery": "^3.3.1",
+    "lodash": "^4.17.5",
+    "ng-recaptcha": "^4.2.1",
+    "ngx-bootstrap": "^3.2.0",
+    "popper.js": "^1.14.7",
+    "rxjs": "^6.4.0",
+    "rxjs-compat": "^6.0.0-rc.0",
+    "tslib": "^1.9.0",
     "web-animations-js": "^2.3.1",
-    "zone.js": "^0.8.18"
+    "zone.js": "^0.8.29"
   },
   "devDependencies": {
-    "@angular/cli": "^1.6.1",
-    "@angular/compiler-cli": "^5.1.1",
-    "@angular/language-service": "^5.1.1",
-    "@types/jasmine": "^2.8.2",
-    "@types/jasminewd2": "^2.0.3",
-    "@types/node": "^8.5.1",
-    "codelyzer": "^4.0.2",
-    "jasmine-core": "^2.8.0",
+    "@angular-devkit/build-angular": "^0.13.5",
+    "@angular/cli": "^7.3.4",
+    "@angular/compiler-cli": "^7.2.8",
+    "@angular/language-service": "^7.2.8",
+    "@types/jasmine": "^2.8.16",
+    "@types/jasminewd2": "^2.0.6",
+    "@types/node": "^8.10.43",
+    "barrelsby": "^2.0.0",
+    "codelyzer": "^4.5.0",
+    "jasmine-core": "^2.99.1",
     "jasmine-spec-reporter": "~4.1.0",
     "karma": "^1.7.1",
     "karma-chrome-launcher": "~2.1.1",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "^1.1.1",
+    "karma-coverage-istanbul-reporter": "^1.4.3",
+    "karma-jasmine": "^1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "node-sass": "^4.7.2",
-    "protractor": "^5.2.2",
-    "ts-node": "^4.0.2",
-    "tslint": "^5.8.0",
-    "typescript": "^2.6.2"
+    "node-sass": "^4.11.0",
+    "protractor": "^5.4.2",
+    "rxjs-tslint": "^0.1.7",
+    "ts-node": "^4.1.0",
+    "tslint": "^5.13.1",
+    "tslint-lines-between-decorator-and-class": "^1.0.1",
+    "typescript": "^3.2.4"
   }
 }
